Patch v0.3
This patch focuses primarily on reworking the German air tree, including reworking the names of all the new Swiss aircraft and weaponry which were added in the Leviathans update. This patch also reworks mid- to top-tier Swedish weaponry, re-adds aircraft and folder names which were forgotten during the port to the new file structure in v0.2, and makes small reworks to various other aircraft in a variety of trees- most notably the JF-17 Thunder, all the MiG-23 Floggers, and all the MiG-21 Fishbeds.