v0.3.1
Patch v0.3.1 for Lord Syrupy's Localization Mod, a UI mod for War Thunder. This patch focuses on re-adding the names of aircraft and folders which were forgotten in the port to the new file structure in v0.2, renames most in-game targeting pods, and makes small fixes to the names of various pieces of weaponry.