v0.3.3
Version 0.3.3 for Lord Syrupy's Localization Mod. This patch primarily focuses on renaming the F-106A, the new variant of the M61A1 Vulcan that got added along with it, and most of the F-16s. It also changes the description of the PIDS modification to be accurate to its function in-game, and continues to rework radar statcards.